refactor + add login post method (returns JWT token)

+++ SpeechGPT.Domain/Request.cs
@@ -5,7 +5,7 @@ namespace SpeechGPT.Domain
 {
     public class Request
     {
-        public BigInteger Id { get; set; }
+        public int Id { get; set; }
         public string Body { get; set; }
         public DateTime CreatedAt { get; set; }
